+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 *  `SetFields` utility that allows to set a field to desired value within a box / circle / sphere [#73](@ref)
 *  Helmholtz Energy equation of state and supporting framework for it for H2 and N2 [#75](@ref)
 *  Time-average and RMS of Vector and Scalar fields [#78]
+*  Added `RotatingWall` velocity boundary condition [#81](@ref)
+
 
 ### Fixed
 * The `UNV3D_mesh` reader has been updated to ensure that the ordering of face nodes is determined in a more robust manner. This resolves some issues when loading a `UNV` mesh that is later used to store simulation results in the `OpenFOAM` format [#64](@ref)